- Fixed Type Casting issue - Reorganized duplicated internal class in the DelimitedDataParser and DelimitedDataParserFactory - Prevented a user from creating an inverted index on a dataset with a variable-length PK
Reviewed-on: http://fulliautomatix.ics.uci.edu:8443/191
Tested-by: Jenkins <jenkins@fulliautomatix.ics.uci.edu>

+++ algebricks/algebricks-runtime/src/main/java/edu/uci/ics/hyracks/algebricks/runtime/base/IUnnestingPositionWriter.java
@@ -3,9 +3,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * you may obtain a copy of the License from
- * 
+ *
  *     http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -18,5 +18,5 @@ import java.io.DataOutput;
 import java.io.IOException;
 
 public interface IUnnestingPositionWriter {
-    public void write(DataOutput dataOutput, int position) throws IOException;
+    public void write(DataOutput dataOutput, long position) throws IOException;
 }
